Updated on 2026-05-25

+++ modules/disp.math.c
@@ -1,12 +1,14 @@
 
 #define _POSIX_C_SOURCE 200809L
-#include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdarg.h>
+#include <stddef.h>
 #include <limits.h>
 #include <ctype.h>
 #include <inttypes.h>
-#include <errno.h>
+#include <stdint.h>
+#include <math.h>
 #ifndef DEBUG
 //#define DEBUG
 #endif
@@ -19,7 +21,7 @@ static disp_box plus_syscall(disp_box *args, int count) {
     if (count == 0) return disp_make_int(0);
 
     int has_double = 0, has_float = 0, has_long = 0;
-    int max_type = FLAG_INT;
+    uint64_t max_type = FLAG_INT;
 
     for (int i = 0; i < count; i++) {
         int t = T(args[i]);
@@ -89,7 +91,7 @@ static disp_box minus_syscall(disp_box *args, int count) {
 
     // First, determine the result type by checking all arguments
     int has_double = 0, has_float = 0, has_long = 0;
-    int max_type = FLAG_INT;
+    uint64_t max_type = FLAG_INT;
     for (int i = 0; i < count; i++) {
         int t = T(args[i]);
         switch (t) {
@@ -154,7 +156,7 @@ static disp_box times_syscall(disp_box *args, int count) {
     if (count == 0) return disp_make_int(1);
 
     int has_double = 0, has_float = 0, has_long = 0;
-    int max_type = FLAG_INT;
+    uint64_t max_type = FLAG_INT;
 
     for (int i = 0; i < count; i++) {
         int t = T(args[i]);
